backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements-base.in
#
--extra-index-url https://pypi.ngc.nvidia.com
--trusted-host pypi.ngc.nvidia.com

annotated-types==0.7.0
    # via pydantic
anyio==4.4.0
    # via starlette
certifi==2024.8.30
    # via requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via uvicorn
colorama==0.4.6
    # via
    #   click
    #   tqdm
fastapi==0.114.2
    # via -r requirements-base.in
filelock==3.16.0
    # via torch
fsspec==2024.9.0
    # via torch
h11==0.14.0
    # via uvicorn
idna==3.9
    # via
    #   anyio
    #   requests
jinja2==3.1.4
    # via torch
llvmlite==0.43.0
    # via numba
markupsafe==2.1.5
    # via jinja2
more-itertools==10.5.0
    # via openai-whisper
mpmath==1.3.0
    # via sympy
networkx==3.3
    # via torch
numba==0.60.0
    # via openai-whisper
numpy==2.0.2
    # via
    #   -r requirements-base.in
    #   numba
    #   openai-whisper
openai-whisper==20231117
    # via -r requirements-base.in
pydantic==2.9.1
    # via
    #   -r requirements-base.in
    #   fastapi
pydantic-core==2.23.3
    # via pydantic
regex==2024.9.11
    # via tiktoken
requests==2.32.3
    # via tiktoken
sniffio==1.3.1
    # via anyio
starlette==0.38.5
    # via fastapi
sympy==1.13.2
    # via torch
tiktoken==0.7.0
    # via openai-whisper
torch==2.4.1
    # via openai-whisper
tqdm==4.66.5
    # via openai-whisper
typing-extensions==4.12.2
    # via
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   torch
urllib3==2.2.3
    # via requests
uvicorn==0.30.6
    # via -r requirements-base.in
websockets==13.0.1
    # via -r requirements-base.in
